getNewData Tasklist operations for reference database now through new DAO.

+++ src/main/java/decodes/tsdb/CTimeSeries.java
@@ -89,7 +89,7 @@ public class CTimeSeries
 	private ArrayList<TimedVariable> vars;
 
 	/** For tracking tasklist records of computations. */
-	private ArrayList<Integer> taskListRecNums; // Note these ARE integers, not DbKey.
+	private ArrayList<Long> taskListRecNums; // Note these ARE integers, not DbKey.
 	
 	/** timeseries  name .used for customised timeseries name in tables and plots **/
 	private String briefDescription;
@@ -137,7 +137,7 @@ public class CTimeSeries
 		vars = new ArrayList<TimedVariable>();
 		unitsAbbr = "unknown";
 		dependentCompIds = null;
-		taskListRecNums = new ArrayList<Integer>();
+		taskListRecNums = new ArrayList<Long>();
 	}
 
 	/**
@@ -701,10 +701,10 @@ Logger.instance().debug3("findInterp(sec=" + sec + ") prevSec=" + prevSec + ", n
 	}
 
 	/** @return the list of tasklist record nums used to populate this ts. */
-	public ArrayList<Integer> getTaskListRecNums() { return taskListRecNums; }
+	public ArrayList<Long> getTaskListRecNums() { return taskListRecNums; }
 
 	/** Adds a recnum to the list of tasklist record nums. */
-	public void addTaskListRecNum(int recnum) { taskListRecNums.add(recnum); }
+	public void addTaskListRecNum(long recnum) { taskListRecNums.add(recnum); }
 
 	/** Clears the list of tasklist record nums. */
 	public void clearTaskListRecNums() { taskListRecNums.clear(); }
